zig default hasher is wyhash, which is quality but expensive