@jinyus 2023-10-12
Old version is actually faster for me (linux):

```
Zig:

Benchmark 1: ./main
Processing time (w/o IO): 15ms
Processing time (w/o IO): 16ms
Processing time (w/o IO): 16ms
Processing time (w/o IO): 16ms
Processing time (w/o IO): 17ms
Processing time (w/o IO): 16ms
Processing time (w/o IO): 15ms
Processing time (w/o IO): 16ms
Processing time (w/o IO): 16ms
Processing time (w/o IO): 16ms
Processing time (w/o IO): 17ms
Processing time (w/o IO): 16ms
Processing time (w/o IO): 16ms
  Time (mean ± σ):      54.6 ms ±   1.3 ms    [User: 43.7 ms, System: 10.9 ms]
  Range (min … max):    52.6 ms …  56.3 ms    10 runs
```

new:

```
Benchmark 1: ./main
Processing time (w/o IO): 21ms
Processing time (w/o IO): 20ms
Processing time (w/o IO): 20ms
Processing time (w/o IO): 20ms
Processing time (w/o IO): 21ms
Processing time (w/o IO): 21ms
Processing time (w/o IO): 20ms
Processing time (w/o IO): 21ms
Processing time (w/o IO): 20ms
Processing time (w/o IO): 21ms
Processing time (w/o IO): 20ms
Processing time (w/o IO): 21ms
Processing time (w/o IO): 20ms
  Time (mean ± σ):      49.8 ms ±   0.8 ms    [User: 45.3 ms, System: 4.5 ms]
  Range (min … max):    48.6 ms …  51.0 ms    10 runs

```