@kaiwu 2023-10-12
I've tried 2 machines, one is faster the other on par, no hurry for the merge anyway, thanks. In Theory FxHash has far less instructions than wyhash, let me check a bit.